requirements.txt
# Python 依赖 / Python dependencies
PyYAML>=6.0

# 系统依赖（需另外安装）/ System dependencies (install separately)
# ─────────────────────────────────────────────────────────────
# ffmpeg        brew install ffmpeg
# Vibe app      https://thewh1teagle.github.io/vibe/
# Claude CLI    npm install -g @anthropic-ai/claude-code
